requirements.txt
fastapi==0.118.0
pandas==2.3.3
playwright==1.57.0
PyQt6==6.9.1
python-dotenv==1.2.1
requests==2.32.5
uvicorn==0.37.0
